added transfer function, added to Makefile to compile Transfer files, and changed data type in savings types to match base class

+++ includes/highInterestSavingsType.h
@@ -11,7 +11,7 @@ public:
 
 
 	double getMinBalance();
-	void withdraw(double amount);
+	void withdraw(int amount);
 	void print();
 
 private:
